backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pypdf==4.2.0
beautifulsoup4==4.12.3
requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1
google-generativeai==0.5.4